app/requirements.txt
Django==3.0.7
gunicorn==20.0.4
psycopg2-binary
django-cors-headers==3.7.0
djangorestframework==3.12.4
